cleaning up code

+++ src/app/api/auth/admin/blogs/route.ts
@@ -2,33 +2,52 @@ import { NextResponse, NextRequest } from "next/server";
 import prisma from "@/lib/prisma";
 
 export async function GET(req: NextRequest) {
-  const blogs = await prisma.blogs.findMany();
-  return NextResponse.json({ status: 200, ok: true, blogs: blogs });
+  try {
+    const blogs = await prisma.blogs.findMany();
+    return NextResponse.json({ status: 200, ok: true, blogs: blogs });
+  } catch (err) {
+    if (err instanceof Error)
+      return NextResponse.json({
+        status: 500,
+        ok: false,
+        message: "Failed to fetch blogs",
+        prisma_error: err.message,
+      });
+  }
 }
 
 export async function DELETE(req: NextRequest) {
   const url = new URL(req.url as string, process.env.NEXT_PUBLIC_CLIENT_URL);
   const blog_id = url.searchParams.get("id");
 
+  if (!blog_id)
+    return NextResponse.json({
+      status: 400,
+      ok: false,
+      message: "Blog id is required",
+    });
+
   try {
-    const blogs = await prisma.blogs.delete({
+    const blog = await prisma.blogs.delete({
       where: {
         blog_id: Number(blog_id),
       },
     });
 
-    if (blogs) {
+    if (blog) {
       return NextResponse.json({
         status: 200,
         ok: true,
-        message: "News successfully deleted!",
+        message: "Blog successfully deleted!",
       });
     }
   } catch (err) {
-    return NextResponse.json({
-      status: 400,
-      ok: false,
-      message: "News failed to delete!",
-    });
+    if (err instanceof Error)
+      return NextResponse.json({
+        status: 400,
+        ok: false,
+        message: "Blog failed to delete!",
+        prisma_error: err.message,
+      });
   }
 }
