+++ src/app/api/auth/admin/users/route.ts
@@ -1,34 +1,53 @@
 import { NextResponse, NextRequest } from "next/server";
 import prisma from "@/lib/prisma";
 
-export async function GET(req: NextRequest, res: NextResponse) {
-  const users = await prisma.users.findMany();
-  return NextResponse.json({ status: 200, ok: true, users: users });
+export async function GET(req: NextRequest) {
+  try {
+    const users = await prisma.users.findMany();
+    return NextResponse.json({ status: 200, ok: true, users: users });
+  } catch (err) {
+    if (err instanceof Error)
+      return NextResponse.json({
+        status: 500,
+        ok: false,
+        message: "Failed to fetch users",
+        prisma_error: err.message,
+      });
+  }
 }
 
 export async function DELETE(req: NextRequest) {
   const url = new URL(req.url as string, process.env.NEXT_PUBLIC_CLIENT_URL);
   const user_id = url.searchParams.get("id");
 
+  if (!user_id)
+    return NextResponse.json({
+      status: 400,
+      ok: false,
+      message: "User id is required",
+    });
+
   try {
-    const users = await prisma.users.delete({
+    const user = await prisma.users.delete({
       where: {
         id: Number(user_id),
       },
     });
 
-    if (users) {
+    if (user) {
       return NextResponse.json({
         status: 200,
         ok: true,
-        message: "News successfully deleted!",
+        message: "User successfully deleted!",
       });
     }
   } catch (err) {
-    return NextResponse.json({
-      status: 400,
-      ok: false,
-      message: "News failed to delete!",
-    });
+    if (err instanceof Error)
+      return NextResponse.json({
+        status: 400,
+        ok: false,
+        message: "User failed to delete!",
+        prisma_error: err.message,
+      });
   }
 }
